GenAI/requirements.txt
transformers
tensorflow
sentence-transformers
accelerate
huggingface_hub
langchain
faiss-cpu
python-dotenv
ipykernel
scikit-learn
tensorboard
nltk
notebook
sentencepiece
jupyter
ipywidgets
python-dotenv
langchainhub
langchain-community
langsmith
langchain_core
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain_openai
chromadb
jupyter-contrib-nbextensions
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain-ollama
streamlit
groq
langchain_groq
langserve[serve]
uvicorn
sse_starlette # this is used for the streaming in the langserve
langchain-chroma
duckduckgo-search
# graphdb requirements  
langchain-neo4j
neo4j
#Finetuning requirements
torch
datasets
peft # this is for the finetuning to work
bitsandbytes
# video summrization requirements
validators # this is streamlit validation
youtube-transcript-api
unstructured # this is for the video summarization to work
# llmmathchain requirement
numexpr # this is for the math chain to work
# astradb pdf rag requirements
cassio # this is for cassandra db
cassandra-driver[uvloop]
tiktoken # this is for openai embeddings, this helps in reducing the token count as we use the embeddings in the retriever and then pass it to the LLM chain.
# pinecone and hybrid search rag
pinecone
pinecone-client
pinecone-text
pinecone-notebooks
#redis db for history with session id
redis
pdfplumber #for tabular data extraction
tabulate #pandas Dataframe to markdown
langchain-huggingface
langchain-openai
python-docx
python-multipart #for file handling
# for Agentic AI
langgraph
pydantic
logfire
langgraph-cli[inmem]
# MCP server
mcp
fastmcp